refactor: consolidate date and workout utility functions
