+++ src/lib/utils/date.ts
@@ -1,5 +1,14 @@
 type WorkoutDateFormat = 'short' | 'full';
 
+export function toLocalDateKey(timestamp: number): string {
+  const date = new Date(timestamp);
+  const year = date.getFullYear();
+  const month = `${date.getMonth() + 1}`.padStart(2, '0');
+  const day = `${date.getDate()}`.padStart(2, '0');
+
+  return `${year}-${month}-${day}`;
+}
+
 export function formatWorkoutDate(
   timestamp: number,
   format: WorkoutDateFormat = 'short'
